backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic[email]>=2.6.0
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
email-validator>=2.0.0
openai>=1.0.0
httpx>=0.25.0



